move copy md script contents in to script
